+++ webinar-api/controllers/webinar.controller.js
@@ -19,3 +19,24 @@ exports.getWebinar = async (req, res) => {
     res.sendStatus(404);
   }
 }
+
+exports.newWebinar = async (req, res) => {
+  const bodyData = Joi.object({
+    link: Joi.string().uri(),
+    title: Joi.string().required(),
+    detail: Joi.string().required()
+  })
+
+  if (bodyData.validate(req.body).error) {
+    res.sendStatus(400);
+    return;
+  }
+
+  try {
+    let webinar = await webinarTable.create(req.body);
+    delete webinar.dataValues.id;
+    res.status(201).send(webinar);
+  } catch (e) {
+    res.sendStatus(500);
+  }
+}
